refactor : UserRole (Enum 파일) 위치변경
