+++ src/main/java/org/example/outsourcing_project/domain/user/service/UserService.java
@@ -1,7 +1,7 @@
 package org.example.outsourcing_project.domain.user.service;
 
 import org.example.outsourcing_project.domain.user.dto.response.UserResponseDto;
-import org.example.outsourcing_project.domain.user.entity.UserRole;
+import org.example.outsourcing_project.domain.user.UserRole;
 
 import java.util.List;
